port api tests to api_tx_approve.go
