+++ services/regulated-assets-approval-server/internal/serve/tx_approve_test.go
@@ -1,0 +1,562 @@
+package serve
+
+import (
+	"context"
+	"net/http"
+	"testing"
+
+	"github.com/stellar/go/amount"
+	"github.com/stellar/go/clients/horizonclient"
+	"github.com/stellar/go/keypair"
+	"github.com/stellar/go/network"
+	"github.com/stellar/go/protocols/horizon"
+	"github.com/stellar/go/protocols/horizon/base"
+	"github.com/stellar/go/services/regulated-assets-approval-server/internal/db/dbtest"
+	"github.com/stellar/go/txnbuild"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestTxApproveHandlerValidate(t *testing.T) {
+	// empty asset issuer KP.
+	h := txApproveHandler{}
+	err := h.validate()
+	require.EqualError(t, err, "issuer keypair cannot be nil")
+	// empty asset code.
+	issuerAccKeyPair := keypair.MustRandom()
+	h = txApproveHandler{
+		issuerKP: issuerAccKeyPair,
+	}
+	err = h.validate()
+	require.EqualError(t, err, "asset code cannot be empty")
+	// No Horizon client.
+	h = txApproveHandler{
+		issuerKP:  issuerAccKeyPair,
+		assetCode: "FOOBAR",
+	}
+	err = h.validate()
+	require.EqualError(t, err, "horizon client cannot be nil")
+	// No network passphrase.
+	horizonMock := horizonclient.MockClient{}
+	h = txApproveHandler{
+		issuerKP:      issuerAccKeyPair,
+		assetCode:     "FOOBAR",
+		horizonClient: &horizonMock,
+	}
+	err = h.validate()
+	require.EqualError(t, err, "network passphrase cannot be empty")
+	// No db.
+	h = txApproveHandler{
+		issuerKP:          issuerAccKeyPair,
+		assetCode:         "FOOBAR",
+		horizonClient:     &horizonMock,
+		networkPassphrase: network.TestNetworkPassphrase,
+	}
+	err = h.validate()
+	require.EqualError(t, err, "database cannot be nil")
+	// Empty kycThreshold.
+	db := dbtest.Open(t)
+	defer db.Close()
+	conn := db.Open()
+	defer conn.Close()
+	h = txApproveHandler{
+		issuerKP:          issuerAccKeyPair,
+		assetCode:         "FOOBAR",
+		horizonClient:     &horizonMock,
+		networkPassphrase: network.TestNetworkPassphrase,
+		db:                conn,
+	}
+	err = h.validate()
+	require.EqualError(t, err, "kyc threshold cannot be less than or equal to zero")
+	// Negative kycThreshold.
+	h = txApproveHandler{
+		issuerKP:          issuerAccKeyPair,
+		assetCode:         "FOOBAR",
+		horizonClient:     &horizonMock,
+		networkPassphrase: network.TestNetworkPassphrase,
+		db:                conn,
+		kycThreshold:      -1,
+	}
+	err = h.validate()
+	require.EqualError(t, err, "kyc threshold cannot be less than or equal to zero")
+	// no baseURL.
+	h = txApproveHandler{
+		issuerKP:          issuerAccKeyPair,
+		assetCode:         "FOOBAR",
+		horizonClient:     &horizonMock,
+		networkPassphrase: network.TestNetworkPassphrase,
+		db:                conn,
+		kycThreshold:      1,
+	}
+	err = h.validate()
+	require.EqualError(t, err, "base url cannot be empty")
+	// Success.
+	h = txApproveHandler{
+		issuerKP:          issuerAccKeyPair,
+		assetCode:         "FOOBAR",
+		horizonClient:     &horizonMock,
+		networkPassphrase: network.TestNetworkPassphrase,
+		db:                conn,
+		kycThreshold:      1,
+		baseURL:           "https://sep8-server.test",
+	}
+	err = h.validate()
+	require.NoError(t, err)
+}
+
+func TestConvertThresholdToReadableString(t *testing.T) {
+	// Prepare raw int64 amountValue.
+	// Context: stellar-core represents asset "amounts" as 64-bit so amounts shown as "500" is represented in stellar-core as 5000000000.
+	var amountValue int64 = 5000000000
+
+	// TEST if no error and if "500.00" returned
+	amountString, err := convertThresholdToReadableString(amountValue)
+	require.NoError(t, err)
+	assert.Equal(t, "500.00", amountString)
+
+	// Prepare amount parsed Int64 from string
+	// Context: env var KYCRequiredPaymentAmountThreshold is the token's unit quantity represented as string.
+	// This string is converted to int64 and passed to the txApproveHandler for payment evaluation.
+	parsedThresholdResult, err := amount.ParseInt64("500")
+
+	// TEST if no error and if "500.00" returned
+	amountString, err = convertThresholdToReadableString(parsedThresholdResult)
+	require.NoError(t, err)
+	assert.Equal(t, "500.00", amountString)
+}
+
+func TestTxApproveHandlerKYCRequiredMessageIfNeeded(t *testing.T) {
+	db := dbtest.Open(t)
+	defer db.Close()
+	conn := db.Open()
+	defer conn.Close()
+
+	// Create tx-approve/ txApproveHandler.
+	issuerAccKeyPair := keypair.MustRandom()
+	horizonMock := horizonclient.MockClient{}
+	kycThresholdAmount, err := amount.ParseInt64("500")
+	require.NoError(t, err)
+	assetGOAT := txnbuild.CreditAsset{
+		Code:   "GOAT",
+		Issuer: issuerAccKeyPair.Address(),
+	}
+	h := txApproveHandler{
+		issuerKP:          issuerAccKeyPair,
+		assetCode:         assetGOAT.GetCode(),
+		horizonClient:     &horizonMock,
+		networkPassphrase: network.TestNetworkPassphrase,
+		db:                conn,
+		kycThreshold:      kycThresholdAmount,
+		baseURL:           "https://sep8-server.test",
+	}
+
+	// TEST if txApproveHandler is valid.
+	err = h.validate()
+	require.NoError(t, err)
+
+	// Preparing payment op for kycRequiredMessageIfNeeded; payment amount is below kycThreshold.
+	destinationKP := keypair.MustRandom()
+	paymentOP := txnbuild.Payment{
+		Destination: destinationKP.Address(),
+		Amount:      "100",
+		Asset:       assetGOAT,
+	}
+
+	// TEST No KYC needed response. actionRequiredMessage should be "".
+	actionRequiredMessage, err := h.kycRequiredMessageIfNeeded(&paymentOP)
+	require.NoError(t, err)
+	require.Empty(t, actionRequiredMessage)
+
+	// Prepare payment op for kycRequiredMessageIfNeeded; payment amount is malformed.
+	paymentOP = txnbuild.Payment{
+		Destination: destinationKP.Address(),
+		Amount:      "ten",
+		Asset:       assetGOAT,
+	}
+
+	// TEST kycRequiredMessageIfNeeded returns error.
+	_, err = h.kycRequiredMessageIfNeeded(&paymentOP)
+	assert.Contains(t,
+		err.Error(),
+		`parsing account payment amount from string to Int64: invalid amount format: ten`,
+	)
+
+	// Preparing payment op for kycRequiredMessageIfNeeded; payment amount is above kycThreshold.
+	paymentOP = txnbuild.Payment{
+		Destination: destinationKP.Address(),
+		Amount:      "501",
+		Asset:       assetGOAT,
+	}
+
+	// TEST Successful KYC required response.
+	// actionRequiredMessage should return "Payments exceeding [kycThreshold] [assetCode] requires KYC approval..." message.
+	actionRequiredMessage, err = h.kycRequiredMessageIfNeeded(&paymentOP)
+	require.NoError(t, err)
+	assert.Equal(t, `Payments exceeding 500.00 GOAT requires KYC approval. Please provide an email address.`, actionRequiredMessage)
+}
+
+func TestTxApproveHandlerHandleKYCRequiredOperationIfNeeded(t *testing.T) {
+	ctx := context.Background()
+	db := dbtest.Open(t)
+	defer db.Close()
+	conn := db.Open()
+	defer conn.Close()
+
+	// Create tx-approve/ txApproveHandler.
+	issuerAccKeyPair := keypair.MustRandom()
+	horizonMock := horizonclient.MockClient{}
+	kycThresholdAmount, err := amount.ParseInt64("500")
+	require.NoError(t, err)
+	assetGOAT := txnbuild.CreditAsset{
+		Code:   "GOAT",
+		Issuer: issuerAccKeyPair.Address(),
+	}
+	h := txApproveHandler{
+		issuerKP:          issuerAccKeyPair,
+		assetCode:         assetGOAT.GetCode(),
+		horizonClient:     &horizonMock,
+		networkPassphrase: network.TestNetworkPassphrase,
+		db:                conn,
+		kycThreshold:      kycThresholdAmount,
+		baseURL:           "https://sep8-server.test",
+	}
+
+	// TEST if txApproveHandler is valid.
+	err = h.validate()
+	require.NoError(t, err)
+
+	// Prepare payment op whose amount is greater than 500 GOATs.
+	sourceKP := keypair.MustRandom()
+	destinationKP := keypair.MustRandom()
+	paymentOP := txnbuild.Payment{
+		SourceAccount: sourceKP.Address(),
+		Destination:   destinationKP.Address(),
+		Amount:        "501",
+		Asset:         assetGOAT,
+	}
+
+	// TEST successful "action_required" response.
+	actionRequiredTxApprovalResponse, err := h.handleKYCRequiredOperationIfNeeded(ctx, sourceKP.Address(), &paymentOP)
+	require.NoError(t, err)
+	wantTXApprovalResponse := txApprovalResponse{
+		Status:       sep8Status("action_required"),
+		Message:      `Payments exceeding 500.00 GOAT requires KYC approval. Please provide an email address.`,
+		StatusCode:   http.StatusOK,
+		ActionURL:    actionRequiredTxApprovalResponse.ActionURL,
+		ActionMethod: "POST",
+		ActionFields: []string{"email_address"},
+	}
+	assert.Equal(t, &wantTXApprovalResponse, actionRequiredTxApprovalResponse)
+
+	// TEST if the kyc attempt was logged in db's accounts_kyc_status table.
+	const q = `
+	SELECT stellar_address
+	FROM accounts_kyc_status
+	WHERE stellar_address = $1
+	`
+	var stellarAddress string
+	err = h.db.QueryRowContext(ctx, q, sourceKP.Address()).Scan(&stellarAddress)
+	require.NoError(t, err)
+	assert.Equal(t, sourceKP.Address(), stellarAddress)
+}
+
+func TestTxApproveHandlerTxApprove(t *testing.T) {
+	ctx := context.Background()
+	db := dbtest.Open(t)
+	defer db.Close()
+	conn := db.Open()
+	defer conn.Close()
+
+	// Perpare accounts on mock horizon.
+	issuerAccKeyPair := keypair.MustRandom()
+	senderAccKP := keypair.MustRandom()
+	receiverAccKP := keypair.MustRandom()
+	assetGOAT := txnbuild.CreditAsset{
+		Code:   "GOAT",
+		Issuer: issuerAccKeyPair.Address(),
+	}
+	horizonMock := horizonclient.MockClient{}
+	horizonMock.
+		On("AccountDetail", horizonclient.AccountRequest{AccountID: issuerAccKeyPair.Address()}).
+		Return(horizon.Account{
+			AccountID: issuerAccKeyPair.Address(),
+			Sequence:  "1",
+			Balances: []horizon.Balance{
+				{
+					Asset:   base.Asset{Code: "ASSET", Issuer: issuerAccKeyPair.Address()},
+					Balance: "0",
+				},
+			},
+		}, nil)
+	horizonMock.
+		On("AccountDetail", horizonclient.AccountRequest{AccountID: senderAccKP.Address()}).
+		Return(horizon.Account{
+			AccountID: senderAccKP.Address(),
+			Sequence:  "2",
+		}, nil)
+	horizonMock.
+		On("AccountDetail", horizonclient.AccountRequest{AccountID: receiverAccKP.Address()}).
+		Return(horizon.Account{
+			AccountID: receiverAccKP.Address(),
+			Sequence:  "3",
+		}, nil)
+
+	// Create tx-approve/ txApproveHandler.
+	kycThresholdAmount, err := amount.ParseInt64("500")
+	require.NoError(t, err)
+	handler := txApproveHandler{
+		issuerKP:          issuerAccKeyPair,
+		assetCode:         assetGOAT.GetCode(),
+		horizonClient:     &horizonMock,
+		networkPassphrase: network.TestNetworkPassphrase,
+		db:                conn,
+		kycThreshold:      kycThresholdAmount,
+		baseURL:           "https://sep8-server.test",
+	}
+
+	// Prepare empty "tx" for txApprove.
+	req := txApproveRequest{
+		Tx: "",
+	}
+
+	// TEST "rejected" response if no transaction is submitted.
+	rejectedResponse, err := handler.txApprove(ctx, req)
+	require.NoError(t, err)
+	wantRejectedResponse := txApprovalResponse{
+		Status:     "rejected",
+		Error:      `Missing parameter "tx".`,
+		StatusCode: http.StatusBadRequest,
+	}
+	assert.Equal(t, &wantRejectedResponse, rejectedResponse)
+
+	// Prepare malformed "tx" for txApprove.
+	req = txApproveRequest{
+		Tx: "BADXDRTRANSACTIONENVELOPE",
+	}
+
+	// TEST "rejected" response if can't parse XDR.
+	rejectedResponse, err = handler.txApprove(ctx, req)
+	require.NoError(t, err)
+	wantRejectedResponse = txApprovalResponse{
+		Status:     "rejected",
+		Error:      `Invalid parameter "tx".`,
+		StatusCode: http.StatusBadRequest,
+	}
+	assert.Equal(t, &wantRejectedResponse, rejectedResponse)
+
+	// Prepare invalid(non generic transaction) "tx" for txApprove.
+	senderAcc, err := handler.horizonClient.AccountDetail(horizonclient.AccountRequest{AccountID: senderAccKP.Address()})
+	require.NoError(t, err)
+	tx, err := txnbuild.NewTransaction(
+		txnbuild.TransactionParams{
+			SourceAccount:        &senderAcc,
+			IncrementSequenceNum: true,
+			Operations: []txnbuild.Operation{
+				&txnbuild.Payment{
+					Destination: receiverAccKP.Address(),
+					Amount:      "1",
+					Asset:       assetGOAT,
+				},
+			},
+			BaseFee:    txnbuild.MinBaseFee,
+			Timebounds: txnbuild.NewInfiniteTimeout(),
+		},
+	)
+	require.NoError(t, err)
+	feeBumpTx, err := txnbuild.NewFeeBumpTransaction(
+		txnbuild.FeeBumpTransactionParams{
+			Inner:      tx,
+			FeeAccount: receiverAccKP.Address(),
+			BaseFee:    2 * txnbuild.MinBaseFee,
+		},
+	)
+	require.NoError(t, err)
+	feeBumpTxEnc, err := feeBumpTx.Base64()
+	require.NoError(t, err)
+	req = txApproveRequest{
+		Tx: feeBumpTxEnc,
+	}
+
+	// TEST "rejected" response if a non generic transaction fails, same result as malformed XDR.
+	rejectedResponse, err = handler.txApprove(ctx, req)
+	require.NoError(t, err)
+	assert.Equal(t, &wantRejectedResponse, rejectedResponse) // wantRejectedResponse is identical to "if can't parse XDR".
+
+	// Prepare transaction sourceAccount the same as the server issuer account for txApprove.
+	issuerAcc, err := handler.horizonClient.AccountDetail(horizonclient.AccountRequest{AccountID: issuerAccKeyPair.Address()})
+	require.NoError(t, err)
+	tx, err = txnbuild.NewTransaction(
+		txnbuild.TransactionParams{
+			SourceAccount:        &issuerAcc,
+			IncrementSequenceNum: true,
+			Operations: []txnbuild.Operation{
+				&txnbuild.Payment{
+					Destination: senderAccKP.Address(),
+					Amount:      "1",
+					Asset:       assetGOAT,
+				},
+			},
+			BaseFee:    txnbuild.MinBaseFee,
+			Timebounds: txnbuild.NewInfiniteTimeout(),
+		},
+	)
+	require.NoError(t, err)
+	txEnc, err := tx.Base64()
+	require.NoError(t, err)
+	req = txApproveRequest{
+		Tx: txEnc,
+	}
+
+	// TEST "rejected" response for sender account; transaction sourceAccount the same as the server issuer account.
+	rejectedResponse, err = handler.txApprove(ctx, req)
+	require.NoError(t, err)
+	wantRejectedResponse = txApprovalResponse{
+		Status:     "rejected",
+		Error:      "The source account is invalid.",
+		StatusCode: http.StatusBadRequest,
+	}
+	assert.Equal(t, &wantRejectedResponse, rejectedResponse)
+
+	// Prepare transaction where transaction's payment operation sourceAccount the same as the server issuer account.
+	tx, err = txnbuild.NewTransaction(
+		txnbuild.TransactionParams{
+			SourceAccount:        &senderAcc,
+			IncrementSequenceNum: true,
+			Operations: []txnbuild.Operation{
+				&txnbuild.Payment{
+					SourceAccount: issuerAccKeyPair.Address(),
+					Destination:   senderAccKP.Address(),
+					Amount:        "1",
+					Asset:         assetGOAT,
+				},
+			},
+			BaseFee:    txnbuild.MinBaseFee,
+			Timebounds: txnbuild.NewInfiniteTimeout(),
+		},
+	)
+	require.NoError(t, err)
+	txEnc, err = tx.Base64()
+	require.NoError(t, err)
+	req = txApproveRequest{
+		Tx: txEnc,
+	}
+
+	// TEST "rejected" response for sender account; payment operation sourceAccount the same as the server issuer account.
+	rejectedResponse, err = handler.txApprove(ctx, req)
+	require.NoError(t, err)
+	wantRejectedResponse = txApprovalResponse{
+		Status:     "rejected",
+		Error:      "There is one or more unauthorized operations in the provided transaction.",
+		StatusCode: http.StatusBadRequest,
+	}
+	assert.Equal(t, &wantRejectedResponse, rejectedResponse)
+
+	// Prepare transaction where operation is not a payment (in this case allowing trust for receiverAccKP).
+	tx, err = txnbuild.NewTransaction(
+		txnbuild.TransactionParams{
+			SourceAccount:        &senderAcc,
+			IncrementSequenceNum: true,
+			Operations: []txnbuild.Operation{
+				&txnbuild.AllowTrust{
+					Trustor:   receiverAccKP.Address(),
+					Type:      assetGOAT,
+					Authorize: true,
+				},
+			},
+			BaseFee:    txnbuild.MinBaseFee,
+			Timebounds: txnbuild.NewInfiniteTimeout(),
+		},
+	)
+	require.NoError(t, err)
+	txEnc, err = tx.Base64()
+	req = txApproveRequest{
+		Tx: txEnc,
+	}
+
+	// TEST "rejected" response if operation is not a payment (in this case allowing trust for receiverAccKP).
+	rejectedResponse, err = handler.txApprove(ctx, req)
+	require.NoError(t, err)
+	wantRejectedResponse = txApprovalResponse{
+		Status:     "rejected",
+		Error:      "There is one or more unauthorized operations in the provided transaction.",
+		StatusCode: http.StatusBadRequest,
+	}
+	assert.Equal(t, &wantRejectedResponse, rejectedResponse)
+
+	// Prepare transaction with multiple operations.
+	tx, err = txnbuild.NewTransaction(
+		txnbuild.TransactionParams{
+			SourceAccount:        &senderAcc,
+			IncrementSequenceNum: true,
+			Operations: []txnbuild.Operation{
+				&txnbuild.Payment{
+					SourceAccount: senderAccKP.Address(),
+					Destination:   receiverAccKP.Address(),
+					Amount:        "1",
+					Asset:         assetGOAT,
+				},
+				&txnbuild.Payment{
+					SourceAccount: senderAccKP.Address(),
+					Destination:   receiverAccKP.Address(),
+					Amount:        "2",
+					Asset:         assetGOAT,
+				},
+			},
+			BaseFee:    txnbuild.MinBaseFee,
+			Timebounds: txnbuild.NewInfiniteTimeout(),
+		},
+	)
+	require.NoError(t, err)
+	txEnc, err = tx.Base64()
+	require.NoError(t, err)
+	req = txApproveRequest{
+		Tx: txEnc,
+	}
+
+	// TEST "rejected" response for sender account; transaction with multiple operations.
+	rejectedResponse, err = handler.txApprove(ctx, req)
+	require.NoError(t, err)
+	wantRejectedResponse = txApprovalResponse{
+		Status:     "rejected",
+		Error:      "Please submit a transaction with exactly one operation of type payment.",
+		StatusCode: http.StatusBadRequest,
+	}
+	assert.Equal(t, &wantRejectedResponse, rejectedResponse)
+
+	// Prepare transaction where sourceAccount seq num too far in the future.
+	tx, err = txnbuild.NewTransaction(
+		txnbuild.TransactionParams{
+			SourceAccount: &horizon.Account{
+				AccountID: senderAccKP.Address(),
+				Sequence:  "50",
+			},
+			IncrementSequenceNum: true,
+			Operations: []txnbuild.Operation{
+				&txnbuild.Payment{
+					SourceAccount: senderAccKP.Address(),
+					Destination:   receiverAccKP.Address(),
+					Amount:        "1",
+					Asset:         assetGOAT,
+				},
+			},
+			BaseFee:    txnbuild.MinBaseFee,
+			Timebounds: txnbuild.NewInfiniteTimeout(),
+		},
+	)
+	require.NoError(t, err)
+	txEnc, err = tx.Base64()
+	require.NoError(t, err)
+	req = txApproveRequest{
+		Tx: txEnc,
+	}
+
+	// TEST "rejected" response if transaction source account seq num is not equal to account sequence+1.
+	rejectedResponse, err = handler.txApprove(ctx, req)
+	require.NoError(t, err)
+	wantRejectedResponse = txApprovalResponse{
+		Status:     "rejected",
+		Error:      "Invalid transaction sequence number.",
+		StatusCode: http.StatusBadRequest,
+	}
+	assert.Equal(t, &wantRejectedResponse, rejectedResponse)
+}
